grocery-intel-demo/app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
databricks-sql-connector==4.0.5
databricks-sdk==0.34.0
